move apply serialization into zmq.serialize
